% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchCorrection.R
\name{batchCorrection}
\alias{batchCorrection}
\title{Wrapper function for scRNA-seq batch effect correction}
\usage{
batchCorrection(X, batchID, method = c("MNN", "harmony"), ...)
}
\arguments{
\item{X}{A matrix of counts (in raw scale), rows are genes, columns are cells}

\item{batchID}{A vector of factors, with length being the number of cells, 
for the batch.}

\item{method}{Cell clustering method to be used}

\item{...}{Other parameters for the corresponding method.}
}
\value{
A batch-corrected matrix. This could be in raw or log scales, depending on the method.
}
\description{
Wrapper function for scRNA-seq batch effect correction
}
