% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellClustering.R
\name{cellClustering}
\alias{cellClustering}
\title{Wrapper function for scRNA-seq cell clustering}
\usage{
cellClustering(
  x,
  method = c("SC3", "Seurat", "monocle", "cidr", "TSCAN", "SHARP"),
  K
)
}
\arguments{
\item{method}{Cell clustering method to be used}

\item{K}{number of clusters, an integer}

\item{X}{A matrix of *normalized* counts (in raw scale), rows are genes, columns are cells}
}
\value{
A vector, with length being the number of cells, for the cluster assignment.
}
\description{
Wrapper function for scRNA-seq cell clustering
}
\note{
The input X should contain all genes. Features are not selected.
}
