% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celltyping.R
\name{celltyping}
\alias{celltyping}
\title{Wrapper function for scRNA-seq celltyping}
\usage{
celltyping(X, method = c("MAST", "SC2P"), ...)
}
\arguments{
\item{X}{A matrix of *raw* counts rows are genes, columns are cells}

\item{method}{celltyping method to be used}

\item{...}{Other parameters for the corresponding method.}
}
\value{
An imputed matrix. This could be in raw or log scales, 
depending on the method.
}
\description{
Wrapper function for scRNA-seq celltyping
}
