% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DE.R
\name{DE}
\alias{DE}
\title{Wrapper function for scRNA-seq DE}
\usage{
DE(X, method = c("MAST", "SC2P"), ...)
}
\arguments{
\item{X}{A matrix of *raw* counts, rows are genes, columns are cells}

\item{method}{DE method to be used}

\item{...}{Other parameters for the corresponding method.}
}
\value{
An imputed matrix. This could be in raw or log scales, 
depending on the method.
}
\description{
Wrapper function for scRNA-seq DE
}
