% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputation.R
\name{imputation}
\alias{imputation}
\title{Wrapper function for scRNA-seq imputation}
\usage{
imputation(X, method = c("DrImpute", "SAVER"), ...)
}
\arguments{
\item{X}{A matrix of *raw* counts, rows are genes, columns are cells}

\item{method}{Imputation method to be used}

\item{...}{Other parameters for the corresponding method.}
}
\value{
An imputed matrix. This could be in raw or log scales, 
depending on the method.
}
\description{
Wrapper function for scRNA-seq imputation
}
