% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{installPkg}
\alias{installPkg}
\title{Install all packages used in this package}
\usage{
installPkg()
}
\description{
The packages used include 
For data normalization: DESeq2, scran, SC2P, SCnorm 
For batch effect correction: batchelor (MNN), harmony
For imputation: SAVER, DrImpute
For cell clustering: SC3, Seurat, monocle3, cidr, TSCAN, SHARP
For differential expression: MAST, SC2P
For cell typing: 
For simulator: POWSC, splatter, scDesign
}
