\name{scRNAseqMethods-package}
\alias{scRNAseqMethods-package}
\alias{scRNAseqMethods}
\docType{package}
\title{
\packageTitle{scRNAseqMethods}
}
\description{
\packageDescription{scRNAseqMethods}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{scRNAseqMethods}
\packageIndices{scRNAseqMethods}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{scRNAseqMethods}

Maintainer: \packageMaintainer{scRNAseqMethods}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
