% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulator.R
\name{simulator}
\alias{simulator}
\title{Wrapper function for scRNA-seq simulator}
\usage{
simulator(
  X,
  ncells,
  totalCounts,
  method = c("scDesign", "splat", "POWSC"),
  ...
)
}
\arguments{
\item{X}{A matrix of *raw* counts for one cell type. Rows are genes, columns are cells.}

\item{ncells}{Number of cells in the simulated data}

\item{totalCounts}{Total read counts for cells. Can be an integer 
(then all cells will have the same counts), or a vector of integers with 
length being the number of cells.}

\item{method}{simulator method to be used}

\item{...}{Other parameters for the corresponding method.}
}
\value{
A simulated count matrix. Rows are genes, columns are cells. 
Number of cells is specified by the user. 
Number of genes is the same as the input.
}
\description{
Given a template dataset, simulate scRNA-seq counts. 
The template data must be from one specific cell types.
}
